requirements-dev.txt
# dev
unittest
flake8
autopep8

Pillow